Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main()
 {
-int num;
+float num;
 scanf("%f",&num);
 printf("You entered: %.2f \n",num);
 return 0;
